+++ src/game/shared/soundenvelope.h
@@ -85,7 +85,7 @@ class ISaveRestoreOps;
 ISaveRestoreOps *GetSoundSaveRestoreOps( );
 
 #define DEFINE_SOUNDPATCH(name) \
-	{ FIELD_CUSTOM, #name, { offsetof(classNameTypedef,name), 0 }, 1, FTYPEDESC_SAVE, NULL, GetSoundSaveRestoreOps( ), NULL }
+	{ FIELD_CUSTOM, #name, { static_cast<int>( offsetof(classNameTypedef,name) ), 0 }, 1, FTYPEDESC_SAVE, NULL, GetSoundSaveRestoreOps( ), NULL }
 
 
 #endif // SOUNDENVELOPE_H
